WIP tests for conda activation PATH management problem

+++ recipe/install-liblief.bat
@@ -1,8 +1,9 @@
 pushd build
 
-ninja -v && ninja -v install
-if %errorlevel% neq 0 exit /b 1
-rmdir /s /q %PREFIX%\share\LIEF\examples
-dir /s /q %PREFIX%\lib
-mkdir %PREFIX%\bin\
-move %PREFIX%\lib\libLIEF.dll %PREFIX%\bin\libLIEF.dll
+  ninja -v && ninja -v install
+  if %errorlevel% neq 0 exit /b 1
+  if exist %PREFIX%\share\LIEF\examples rmdir /s /q %PREFIX%\share\LIEF\examples
+  if not exist %PREFIX%\bin mkdir %PREFIX%\bin\
+  if not exist %PREFIX%\bin\libLIEF.dll move %PREFIX%\lib\libLIEF.dll %PREFIX%\bin\libLIEF.dll
+
+popd
